requirements.txt
pytz
Flask
Flask-MySQLdb
python-dotenv
Flask-Login
Flask-WTF
Werkzeug